% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Evans}
\alias{Evans}
\title{Evans County dataset}
\format{
A data frame with 609 rows and 9 variables:
\describe{
\item{CDH}{outcome variable; 1 = coronarty disease}
\item{CAT}{1 = high, 0 = normal catecholamine level}
\item{AGE}{age (in years)}
\item{CHL}{cholesterol, mg/dl}
\item{SMK}{1 = subject has ever smoked}
\item{ECG}{1 = presence of electrocardiogram abnormality}
\item{DBP}{diastolic blood pressure, mmHg}
\item{SBP}{systolic blood pressure, mmHg}
\item{HPT}{1 = SBP greater than or equal to 160 or DBP greater than or equal to 95}
}
}
\usage{
Evans
}
\description{
Data from cohort study in which white males in Evans County were followed for 7 years,
with coronary heart disease as the outcome of interest. This dataset comes the
\code{lbref} package.
}
\keyword{datasets}
